services/requirement.txt
uvicorn
pydantic
python-dotenv
psycopg2-binary
sqlalchemy
pwdlib[argon2]   
python-jose